requirements.txt
fastapi
uvicorn
whisper
torchaudio
scikit-learn
speechbrain
nltk
numpy
spacytextblob
spacy
nltk
librosa
pyloudnorm
matplotlib
groq
python-dotenv
python-multipart


